新しいフォルダ構成案に合わせて既存ファイル移動
