+++ frontend-nextjs/app/(auth)/signin/page.tsx
@@ -1,0 +1,12 @@
+import type React from "react";
+import { SignInForm } from "features/signin/SignInForm";
+
+const SignInPage: React.FC = () => {
+  return (
+    <main>
+      <SignInForm />
+    </main>
+  );
+};
+
+export default SignInPage;
