--- frontend-nextjs/app/signin/page.tsx
@@ -1,12 +1,0 @@
-import type React from "react";
-import { SignInForm } from "../components/signin/SignInForm";
-
-const SignInPage: React.FC = () => {
-	return (
-		<main>
-			<SignInForm />
-		</main>
-	);
-};
-
-export default SignInPage;
