+++ frontend-nextjs/lib/firebase.ts
